refactor: extract account-deletion steps into named functions
verifyAndDeleteUser was a single ~600-line function with 13 manually
ordered delete/update steps and FK-ordering constraints documented only
as inline comments. Extract each step into a small named function in a
new account-deletion-steps.ts sibling module, called in the same order
from the same transaction. Comments describing ordering constraints move
with their step; no query changes, no reordering.

Each step that deletes, clears, or reassigns a user reference also
declares the tables it covers, so a coverage guard can be built from
those declarations instead of a hand-maintained parallel list.
